feat(lint): move lint flows out of scripts-to-lint
Separate lint flows allow multiple sources to be linted, and removes the
dependency between the lint tasks and `scripts-to-build`. This commit
introduces a separate lint flow for linting normal JavaScript files, and
JavaScript specs.

Breaks:
- the `globs.source.test` glob, previously a collection of glob patterns,
will be expected to be an object, with a `js` property containing the
collection of globs for `js` spec files.
- Without the `spell-karma` plugin, and the relevant update containing
the negating glob for `*.spec.js` in `globs.source.js`, `*.spec.js` files
will also be included in the build. If needed, manually negate `*.spec.js`
for `globs.source.js` in `warlock.json`.