fix(util): fix _isVendor name collisions
Match the file names including the preceding `/` to account for similar files names. For example,
`Restangular.js` would collide with `angular.js`, but `/Restangular.js` will not collide with
`/angular.js`.

This fixes most cases of name collision, unless there are two files with exact same names from
different libraries. This is much less likely, but not impossible.

There must be a better way of checking if a reference is to a vendor file!